requirements.txt
ago==0.0.95
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blis==0.7.11
boto3==1.28.77
botocore==1.31.77
bs4==0.0.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
cymem==2.0.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-cors-headers==4.3.0
django-import-export==3.3.1
djangorestframework==3.14.0
docx2txt==0.8
dotmap==1.3.30
elastic-transport==8.10.0
elasticsearch==8.10.1
emoji==1.7.0
et-xmlfile==1.1.0
faust-cchardet==2.1.19
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
frozenlist==1.4.0
ftfy==6.1.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hjson==3.1.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hurry.filesize==0.9
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jellyfish==1.0.1
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
langcodes==3.3.0
langdetect==1.0.9
libretranslatepy==2.1.1
locationtagger==0.0.1
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.10
networkx==3.2.1
news-please==1.5.35
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.1
odfpy==1.4.1
openai==0.28.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
parsel==1.8.1
Pillow==10.1.0
plac==1.4.1
preshed==3.0.9
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
pygoogletranslation==2.0.6
PyMySQL==1.1.0
pyOpenSSL==23.3.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
readability-lxml==0.8.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
rfc3986==1.5.0
s3transfer==0.7.0
Scrapy==2.11.0
segtok==1.5.11
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
tablib==3.5.0
tabulate==0.9.0
textblob==0.17.1
thinc==8.2.1
tinysegmenter==0.3
tldextract==5.0.1
tqdm==4.66.1
Twisted==22.10.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.7
w3lib==2.1.2
warcio==1.7.4
wasabi==1.1.2
wcwidth==0.2.9
weasel==0.3.3
xlrd==2.0.1
xlwt==1.3.0
yake @ git+https://github.com/LIAAD/yake@238ae58c5ba39326a96862ee0e9cb817e5958440
yarl==1.9.2
zope.interface==6.1
undetected-chromedriver==3.5.3
webdriver-manager==4.0.1
selenium==4.13.0
PyVirtualDisplay==3.0
PyAutoGUI==0.9.54
fake-useragent==1.3.0

yake~=0.4.8